# src/main/java/com/slotsapi/dto/responses/IntervalResponseDto.java
package com.slotsapi.dto;

import lombok.Getter;
import lombok.Setter;
import java.time.DayOfWeek;
import java.time.LocalTime;

@Getter
@Setter
public class IntervalResponseDto {
    private Long id;
    private DayOfWeek dayOfWeek;
    private LocalTime startTime;
    private LocalTime endTime;
}